Final submission
Please note, Spotify is not supported in South Africa. I have had to try use an alternative API from Deezer. However, I was unable to achieve the desired result. Please take note of this while moderating. I have included the Deezer methods for revision. The current submission is purely based off the CodeCademy examples as I was unable to test.